+++ src/main/java/net/mcreator/ragemod/procedures/YellowGlowingVineBlockValidPlacementConditionProcedure.java
@@ -36,9 +36,9 @@ public class YellowGlowingVineBlockValidPlacementConditionProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if ((world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getMaterial() == net.minecraft.block.material.Material.LEAVES
+		if ((world.getBlockState(new BlockPos(x, y + 1, z))).getMaterial() == net.minecraft.block.material.Material.LEAVES
 				|| BlockTags.getCollection().getTagByID(new ResourceLocation("minecraft:leaves"))
-						.contains((world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z))).getBlock())) {
+						.contains((world.getBlockState(new BlockPos(x, y + 1, z))).getBlock())) {
 			return true;
 		}
 		return false;
